невозможно удалить счёт на котором висят фин.цели
git-svn-id: svn://hm.podzone.net/trunk@2029 240b7755-b59a-4df0-b20f-7dab6991996d

+++ models/accounts.model.php
@@ -313,21 +313,26 @@ class Accounts_Model
      */
     public function deleteAccount($id)
     {
-        
+        $fin = new Targets_Model();
+        $noFinTarget = $fin->countTargetsOnAccount($id);// если ноль значит удаляем
+        if (!$noFinTarget){
 
-	$sql = "DELETE FROM
-                    account_field_values
-                WHERE
-                    account_fieldsaccount_field_id = ?;";
-        $this->db->query($sql, $id);
+            $sql = "DELETE FROM
+                        account_field_values
+                    WHERE
+                        account_fieldsaccount_field_id = ?;";
+            $this->db->query($sql, $id);
 
-        $sql = "DELETE FROM accounts WHERE `account_id` = ? and `user_id` = ?;";
-        $this->db->query($sql, $id, $this->user_id);//удаляем счёт.
-        /*$sql = "DELETE FROM operation WHERE `account_id` = ? and `user_id` = ?;";
-        $this->db->query($sql, $id, $this->user_id);//удаляем операции со счётом.*/
-        Core::getInstance()->user->initUserAccounts();
-        Core::getInstance()->user->save();
-        return true;
+            $sql = "DELETE FROM accounts WHERE `account_id` = ? and `user_id` = ?;";
+                $this->db->query($sql, $id, $this->user_id);//удаляем счёт.
+            /*$sql = "DELETE FROM operation WHERE `account_id` = ? and `user_id` = ?;";
+            $this->db->query($sql, $id, $this->user_id);//удаляем операции со счётом.*/
+            Core::getInstance()->user->initUserAccounts();
+            Core::getInstance()->user->save();
+            return true;
+        }
+        else
+            return 'cel';
     }
     /**
      * Функция которая отсылает список счетов
